requirements.txt
numpy==1.25.2
opencv-python==4.8.0.74
Pillow==10.0.0
tk==0.1.0
